Add notes support
