+++ packages/gatsby-theme-next/decks/hello.mdx
@@ -1,10 +1,17 @@
-import { Steps } from '../src'
+import { Steps, Notes } from '../src'
 
 # Hello
 
 ---
 
 `@mdx-deck/gatsby-theme-next`
+
+<Notes>
+
+Hi
+Hello, my secret notes...
+
+</Notes>
 
 ---
