+++ packages/gatsby-theme-next/src/components/notes.js
@@ -1,0 +1,15 @@
+import React, { useEffect } from 'react'
+import useDeck from '../hooks/use-deck'
+
+export const Notes = props => {
+  const context = useDeck()
+  useEffect(() => {
+    context.register(context.index, {
+      notes: props.children,
+    })
+  }, [props.children])
+
+  return false
+}
+
+export default Notes
